+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Show both unified and transparent addresses in Simple view Receive dialog ([#113](https://github.com/LeakIX/zcash-web-wallet/issues/113), [#114](https://github.com/LeakIX/zcash-web-wallet/pull/114))
+
 ## [0.1.0] - 2025-12-30
 
 ### Added